Move base64 to util/
