+++ AddressBook.h
@@ -8,7 +8,7 @@
 #include <iostream>
 #include <mutex>
 #include <boost/asio.hpp>
-#include "base64.h"
+#include "util/base64.h"
 #include "util/util.h"
 #include "Identity.h"
 #include "Log.h"
