+++ I2NPProtocol.cpp
@@ -7,7 +7,6 @@
 #include "RouterContext.h"
 #include "NetDb.h"
 #include "Tunnel.h"
-#include "base64.h"
 #include "transport/Transports.h"
 #include "Garlic.h"
 #include "I2NPProtocol.h"
