+++ I2PTunnel.cpp
@@ -1,5 +1,4 @@
 #include <cassert>
-#include "base64.h"
 #include "Log.h"
 #include "Destination.h"
 #include "ClientContext.h"
