+++ filelist.mk
@@ -1,12 +1,12 @@
 COMMON_SRC = \
   transport/NTCPSession.cpp transport/SSU.cpp transport/SSUSession.cpp \
   transport/SSUData.cpp transport/Transports.cpp \
-  util/util.cpp \
+  util/util.cpp util/base64.cpp \
   CryptoConst.cpp Datagram.cpp Garlic.cpp I2NPProtocol.cpp LeaseSet.cpp \
   Log.cpp NetDb.cpp NetDbRequests.cpp Profiling.cpp Reseed.cpp \
   RouterContext.cpp RouterInfo.cpp Signature.cpp Streaming.cpp Identity.cpp \
   TransitTunnel.cpp Tunnel.cpp TunnelEndpoint.cpp TunnelPool.cpp \
-  TunnelGateway.cpp Destination.cpp UPnP.cpp aes.cpp base64.cpp
+  TunnelGateway.cpp Destination.cpp UPnP.cpp aes.cpp
 
 
 ifeq ($(UNAME),Darwin)
