+++ transport/NTCPSession.cpp
@@ -3,9 +3,9 @@
 #include "I2PEndian.h"
 #include <cryptopp/dh.h>
 #include <cryptopp/adler32.h>
-#include "base64.h"
+#include "util/base64.h"
 #include "Log.h"
-#include "Timestamp.h"
+#include "util/Timestamp.h"
 #include "CryptoConst.h"
 #include "I2NPProtocol.h"
 #include "RouterContext.h"
